Refactored so that we can create a tx-only switch loop tester. This will be useful for integrating a new FSK transmitter without receiver
