+++ src/legacy.py
@@ -145,6 +145,5 @@ if __name__ == "__main__":
     finally:
         radio.finished()
 
-
 # END
